refactor(event/common): rename exports/imports
