+++ src/event/common/constants.ts
