+++ src/event/listener/cache.ts
@@ -5,7 +5,7 @@ import {log} from '@jovijovi/pedrojs-common';
 import {Cache} from '../../common/cache';
 import {customConfig} from '../../config';
 import {DefaultDumpCacheDir, DefaultDumpCacheFile, DefaultNameOfContractOwnerCache,} from './params';
-import {TimeDayInMs, TimeHourInMs} from '../common';
+import {TimeDayInMs, TimeHourInMs} from '../common/constants';
 
 // LoadCacheFromFile load cache from file
 export function LoadCacheFromFile(): number {
