+++ src/event/listener/event.ts
@@ -5,8 +5,8 @@ import {log, util} from '@jovijovi/pedrojs-common';
 import {network} from '@jovijovi/ether-network';
 import {customConfig} from '../../config';
 import {DefaultCallbackJobConcurrency, DefaultDumpCacheInterval, DefaultLoopInterval} from './params';
-import {EventNameTransfer, TimeSecondInMs} from '../common';
-import {EventTransfer, Response} from './types';
+import {EventNameTransfer, TimeSecondInMs} from '../common/constants';
+import {EventTransfer, Response} from '../common/types';
 import {GetContractOwner} from './abi';
 import {DumpCacheToFile, GetContractOwnerCacheConfig, InitCache, LoadCacheFromFile} from './cache';
 import {CheckEventType, CheckTopics} from '../utils';
